+++ lib/features/number_trivia/domain/entities/number_trivia.dart
@@ -10,7 +10,7 @@ import 'package:equatable/equatable.dart';
 class NumberTrivia extends Equatable{
 
   final String text;
-  final int number;
+  final double number;
 
   const NumberTrivia({
      required this.text,
