+++ lib/features/number_trivia/domain/repositories/number_trivia_repository.dart
@@ -6,7 +6,7 @@ import '../../../../core/error/failures.dart';
 
 abstract class NumberTriviaRepository {
 
-  Future<Either<Failure, NumberTrivia>> getConcreteNumberTrivia(int? number);
+  Future<Either<Failure, NumberTrivia>> getConcreteNumberTrivia(double? number);
   Future<Either<Failure, NumberTrivia>> getRandomNumberTrivia();
 
 }
